+++ assets/js/components/Login.jsx
@@ -4,7 +4,7 @@ import TrentoLogo from '@static/trento.svg';
 import { useDispatch, useSelector } from 'react-redux';
 import { toast } from 'react-hot-toast';
 import { useNavigate, useSearchParams } from 'react-router-dom';
-import { performLogin } from '@state/actions/auth';
+import { performLogin } from '@state/auth';
 import classNames from 'classnames';
 
 export default function Login() {
